# refactor(CSRangeHistogram): Refactor csrangehistogram builders